requirements.txt
# Interview AI Studio - Requirements for FastAPI Version

# --- Core Framework ---
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<0.30

# --- NLP & ML ---
numpy>=1.24,<1.27
scikit-learn>=1.3,<1.4
scipy>=1.10,<1.12
transformers[sentencepiece]>=4.30,<4.36
torch>=2.0,<2.2
sentence-transformers>=2.2,<2.3

# --- Resume Parsing ---
python-docx>=1.1,<1.2
PyMuPDF>=1.23,<1.24  # import as `fitz`

# --- Transcription ---
openai-whisper>=20231106

# --- Video Analysis ---
mediapipe>=0.10,<0.11
opencv-python>=4.8,<4.9

# --- PDF/Docx Processing & Utilities ---
Pillow>=10.0,<10.2
python-dotenv>=1.0,<1.1
ffmpeg-python>=0.2,<0.3  # FFmpeg must be installed separately system-wide

# --- Optional Frontend (Streamlit) ---
streamlit>=1.30,<2.0

# --- Dev Tools (Optional) ---
# black
# isort
# ipython
